+++ docs/changelog.md
@@ -4,6 +4,10 @@ ITables ChangeLog
 1.4.5-dev (2023-01-??)
 ------------------
 
+**Fixed**
+- Fixed an issue when `lengthMenu` is a 2D array ([#151](https://github.com/mwouts/itables/issues/151))
+
+
 **Changed**
 - We make sure that no argument passed to `show` is equal to `None` (for most options this results in a datatable not loading)
 - Running the test collection will not update the CSV files used for testing anymore